requirements/base.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Celery
celery==5.3.4
redis==5.0.1

# File Processing
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.2.0

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# AWS S3 (for production)
boto3==1.34.23
django-storages==1.14.2

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# API Documentation
drf-spectacular==0.27.0
